moved upgrade appsmith to manual suite
